requirements.txt
Flask==3.1.2
flask-cors==6.0.2
ollama==0.6.1
perplexityai==0.22.0
pydantic==2.12.5
httpx==0.28.1
Werkzeug==3.1.4
